Restores Windows PowerShell 5.1 compatibility.
Versions 3.0.0 to 3.1.0 (inclusive)  of this module contained a bug that
prevented it from working correctly on Windows PowerShell 5.1.  The call
to `ConvertFrom-SecureString` used the parameter `-AsPlainText`, which
does not exist before PowerShell 7.0.

This commit fixes that by checking to see if `ConvertFrom-SecureString`
supports the `-AsPlainText` parameter.  If so, it is called with it, and
if not, it's called without it.

Fixes #7